Validate file source file storage request separate from resolver
